Connectivity

+++ RecycleRush/MatchAnalysis/CalculateTeamStats.m
@@ -45,7 +45,7 @@
     NSPredicate *pred = [NSPredicate predicateWithFormat:@"tournamentName = %@ AND results = %@ AND (matchType = %@ || matchType = %@)", tournament, [NSNumber numberWithBool:YES], [EnumerationDictionary getValueFromKey:@"Qualification" forDictionary:matchTypeDictionary], [EnumerationDictionary getValueFromKey:@"Elimination" forDictionary:matchTypeDictionary]];
     NSArray *matches = [allMatches filteredArrayUsingPredicate:pred];
 
-    int numberOfMatches = [matches count];
+    NSUInteger numberOfMatches = [matches count];
     for (int j=1; j<[parameterList count]; j++) {
         NSDictionary *parameter = [parameterList objectAtIndex:j];
         float total = 0.0;
